+++ src/cachelib/uwsgi.py
@@ -55,17 +55,17 @@ class UWSGICache(BaseCache):
         return bool(self._uwsgi.cache_del(key, self.cache))
 
     def set(
-        self, key: str, value: _t.Any, timeout: _t.Optional[int] = None
-    ) -> _t.Optional[bool]:
+        self, key: str, value: _t.Any, timeout: int | None = None
+    ) -> _t.Literal[True] | None:
         result = self._uwsgi.cache_update(
             key,
             self.serializer.dumps(value),
             self._normalize_timeout(timeout),
             self.cache,
-        )  # type: bool
+        )  # type: _t.Optional[_t.Literal[True]]
         return result
 
-    def add(self, key: str, value: _t.Any, timeout: _t.Optional[int] = None) -> bool:
+    def add(self, key: str, value: _t.Any, timeout: int | None = None) -> bool:
         return bool(
             self._uwsgi.cache_set(
                 key,