+++ tests/test_interface_uniformity.py
@@ -13,12 +13,13 @@ from cachelib import SimpleCache
 @pytest.fixture(autouse=True)
 def create_cache_list(request, tmpdir):
     mc = MemcachedCache(servers=["127.0.0.1:11212"])
-    mc._client.flush_all()
+    mc.clear()
     rc = RedisCache(port=6360)
     rc._write_client.flushdb()
     request.cls.cache_list = [FileSystemCache(tmpdir), mc, rc, SimpleCache()]
 
 
+@pytest.mark.network
 @pytest.mark.usefixtures("redis_server", "memcached_server")
 class TestInterfaceUniformity:
     def test_types_have_all_base_methods(self):